Moved all validations to validations.py
