+++ qiskit_ibm_runtime/base_runtime_job.py
@@ -30,8 +30,7 @@ from qiskit.providers.jobstatus import JobStatus as RuntimeJobStatus
 
 from qiskit_ibm_runtime import qiskit_runtime_service
 
-from .utils import utc_to_local
-from .utils.utils import validate_job_tags
+from .utils import utc_to_local, validate_job_tags
 from .utils.queueinfo import QueueInfo
 from .utils.deprecation import issue_deprecation_msg
 from .constants import DEFAULT_DECODERS, API_TO_JOB_ERROR_MESSAGE
