+++ qiskit_ibm_runtime/ibm_backend.py
@@ -45,7 +45,6 @@ from .provider_session import (
     Session as ProviderSession,
 )
 
-from .utils.utils import validate_job_tags
 from . import qiskit_runtime_service  # pylint: disable=unused-import,cyclic-import
 from .runtime_job import RuntimeJob
 
@@ -60,7 +59,7 @@ from .utils.backend_decoder import (
 from .utils.deprecation import issue_deprecation_msg
 from .utils.options import QASM2Options, QASM3Options
 from .api.exceptions import RequestsApiError
-from .utils import local_to_utc, are_circuits_dynamic
+from .utils import local_to_utc, are_circuits_dynamic, validate_job_tags
 
 from .utils.pubsub import Publisher
