+++ qiskit_ibm_runtime/utils/__init__.py
@@ -24,7 +24,14 @@ from .utils import (
     get_runtime_api_base_url,
     resolve_crn,
     are_circuits_dynamic,
-    validate_no_dd_with_dynamic_circuits,
 )
+from .validations import (
+    validate_estimator_pubs,
+    validate_classical_registers,
+    validate_no_dd_with_dynamic_circuits,
+    validate_isa_circuits,
+    validate_job_tags,
+)
+
 from .json import RuntimeEncoder, RuntimeDecoder, to_base64_string
 from . import pubsub
